# Refactor use_heuristics for fast filter evaluation