@andrews05 2022-11-20
One other thought I had: It might be good to expose the use_heuristics option in the cli, and maybe rename it to something that better reflects what it does here. `--fast-filter-evaluation` seems a bit much though 😅. Thoughts?