@shssoichiro 2022-12-01
Thanks, other than the one nit I think this looks good! I don't mind the addition of a CLI option for it as well.